refactor: move note component
