+++ src/app/components/note/note.component.html
