+++ src/app/components/note/note.component.ts
